HeliosSoloDeployment: support Google Container Registry credentials
If the configuration profile used by helios-solo contains an entry for
`google-container-registry.credentials`, then HeliosSoloDeployment will
launch the helios-solo container with additional configuration:

- add an environment variable for `HELIOS_AGENT_OPTS` which contains the
flag `--docker-gcp-account-credentials=<file>`
- add a volume bind for the directory containing the credentials file so
that helios agent can find the credentials pointed to above

The purpose of these checks is to allow helios-agent inside the
helios-solo container to be able to fetch images from Google Container
Registry for any tests that need to do so.

If the value is not set in the configuration file for
helios-testing/helios-solo (`helios.conf` or `helios-base.conf`), or to
override the value, the system property
`-Dhelios.testing.profiles.local.google-container-registry.credentials`
can be set to the file path.

If the configuration entry points to a file that doesn't exist then it
is ignored.