client: Support authentication using access tokens
This adds support for acquiring an access token within the HeliosClient
Builder via Google Application Default Credentials. When available, the
client will prefer to send an access token in requests as the means for
authentication. As we don't intend for Google credentials to be a hard
requirement, the client will fall back onto the existing authentication
mechanism of self-signed certs and mutual TLS if tokens are unavailable.