helios-testing: shade and relocate guava classes
Adds a second artifact for helios-testing (with classifier=`shaded`)
that relocates Guava classes inside the JAR to a new package name.

This aims to cut down on pain internally whenever a team wants to
upgrade a project to a newer version of helios-testing, as Guava
incompatibilities from different versions being used is a common
headache or roadblock.

Closes #1119