tools: Finalize google-credentials flag
This commit cleans up the Google credentials related flags by defining a
single flag to enable/disable use of Google Credentials in the client,
and an environment variable to control the path to a particular
credentials file. As with the `spotify/docker-client` repository, we
check for our own environment variable (`HELIOS_GOOGLE_CREDENTIALS`)
rather than use the standard `GOOGLE_APPLICATION_CREDENTIALS` in order
to avoid wrongly influencing other tools.