client: Support authentication using access tokens (#1135)
* client: Support authentication using access tokens

This adds support for acquiring an access token within the HeliosClient
Builder via Google Application Default Credentials. When available, the
client will prefer to send an access token in requests as the means for
authentication. As we don't intend for Google credentials to be a hard
requirement, the client will fall back onto the existing authentication
mechanism of self-signed certs and mutual TLS if tokens are unavailable.

* tools: Add option to disable using Google ADC

* client: Add Authorization header additively

This commit adds support within the AuthenticatingHttpConnector to add
an Authorization header when possible while still going through the
ssh-agent or certificate file based connection mechanism. Supporting
both the token and certificate methods of authentication simultaneously
allows the backend to be upgraded independently from the client.

* tools: Add flags controlling use of Google Credentials

* tools: Finalize google-credentials flag

This commit cleans up the Google credentials related flags by defining a
single flag to enable/disable use of Google Credentials in the client,
and an environment variable to control the path to a particular
credentials file. As with the `spotify/docker-client` repository, we
check for our own environment variable (`HELIOS_GOOGLE_CREDENTIALS`)
rather than use the standard `GOOGLE_APPLICATION_CREDENTIALS` in order
to avoid wrongly influencing other tools.

* docs: Document auth via Google access tokens