Add log message for rolling-update failures (#1157)
Presently, helios-master will emit Kafka and Google Cloud PubSub events
when a rolling-update fails (in addition to storing the failure and
error message in ZooKeeper). This change adds a log message for the
failure as well, which should be useful for systems without a defined
event handler or for quick-and-dirty integrations that don't need the
larger feature set of something like Kafka or Google Cloud PubSub.