log more details when rolling-update times out
Log the elapsed time of a deploy when it has timed out, to help in
debugging problems where rollouts appear to timeout too fast.

Ideally the elapsed time could be included in the error message reported
back to the user, but this seemed like too large of a code change, as
the elapsed time is internal to this `isRolloutTimedOut` function.