agent: Report when container exits in healthchecks (#1202)
The `TaskMonitor` has logic to throttle a container if the container
restarts too frequently. Prior to this change, however, the throttling
logic is bypassed when a container exits during healthchecks, since in
that case we throw an exception. The `TaskRunner` should report the
container's exit, even if the container never reached "running".