cli: refresh Google access token if needed
by using `GoogleCredentials.getRequestMetadata()`.

[OAuth2Credentials](https://github.com/google/google-auth-library-java/blob/master/oauth2_http/java/com/google/auth/oauth2/OAuth2Credentials.java)
has methods `refresh()`. But there's no method to check if the token
is close to expiring and refresh if so. I want to do
[these few lines](https://github.com/google/google-auth-library-java/blob/a8929df490d3751f2349986264290a074dee35e1/oauth2_http/java/com/google/auth/oauth2/OAuth2Credentials.java#L160-L162), but these methods are private.

My use case is that I have a long-lived `GoogleCredentials` instance.
I only want to refresh when I need to to minimize unnecessary overhead in making
requests to the backend. There are the methods
[`getRequestMetadata()`](https://github.com/google/google-auth-library-java/blob/a8929df490d3751f2349986264290a074dee35e1/oauth2_http/java/com/google/auth/oauth2/OAuth2Credentials.java#L158) which will essentially do the check and refresh logic. So we use this.

See https://github.com/google/google-auth-library-java/issues/147.