remove GoogleCredentialsAccessTokenSupplier.enabled field
instead, if the use of the google tokens in HeliosClient.Builder has
been disabled, then HeliosClient.Builder should configure itself to use
an AuthorizationHeaderSupplier that always returns absent Optionals -
this logic does not need to be in the GoogleCredentials class.

Also remove the options in the Builder for setting a static Google
Access Token (seems unneeded - and can be done by passing in a
customized instance of GoogleCredentialsAccessTokenSupplier to the
Builder) and the option to set the scopes used (same as previous point).