auth: throw exception if file at HELIOS_GOOGLE_CREDENTIALS cannot be read
This changes the behavior when the HELIOS_GOOGLE_CREDENTIALS environment
variable is set to a file that does not exist or cannot be read from
falling to Application Default Credentials to instead throw an
exception.

If the environment variable contains invalid configuration, it seems
better to throw an exception rather than silently use a different set of
(possibly incorrect) credentials.

As a part of this, I moved the GoogleCredential-loading logic into
HeliosClient.Builder to simplify GoogleCredentialsAccessTokenSupplier
even further.