client: fix wrong call to google-auth-library-java
We were mistakenly calling `GoogleCredentials.refreshAccessToken()`
which always makes a call to retrieve a new access token regardless of expiration.
In addition, this method returns an `AccessToken` which we accidentally
ignored. But this method but never sets the `AccessToken`
for the `getAccessToken()` method which we subsequently called.

Instead, we should use `GoogleCredentials.refreshIfExpired()`
which refreshes the token if only if it's close to expiring and returns
void. Our next call to `getAccessToken()` will actually return the refreshed
token in this case.

See https://github.com/google/google-auth-library-java/pull/163