drop target=1.7 for client, testing and tools modules
helios-client, -testing and -tools are currently built with a 1.7
target. In 0bf718b1a these modules were excluded from being upgraded to
a 1.8 target to not exclude people who still used 1.7 or depended on it
(with the CLI, or Hadoop-using teams that were stuck on 1.7).

It seems like we can drop this now.