remove usage of deprecated Futures.catching from Guava
The overload of this method that does not take an Executor as an
argument is removed from Guava starting in release 26.0.

For anyone using helios-testing in a project that uses Guava >= 26.0,
this code in helios-testing causes NoSuchMethodErrors at runtime.

This is related to #1257 which I think missing this `Futures` method.