Fix NPE in OldJobReaper
The primary purpose of this change is to avoid extraneous NPE errors in the
warning logs. Other than that, this PR won't affect any the helios master in
any meaningful way.

OldJobReaper periodically garbage collects old jobs that haven't been run in a
certain amount of time. It gets a list of all jobs, and slowly iterates through
each one checking when it was last run. By the time it checks, the job may have
already been deleted. This causes noisy NPE errors in the logs.