master: log more info on ZK error when rolling-update fails
We log the stacktrace of KeeperException. But this class
extends Exception and doesn't override `toString()`.
So right now the error message and stacktrace don't state the
ZK path, if any, that the exception occurred on. Knowing the ZK path is helpful
in debugging rolling-update problems.

`KeeperException.getMessage()` returns the error code and ZK path if present.